Unifying some test cases
