+++ src/ecdsa/ot_based_ecdsa/triples/correlated_ot_extension.rs
@@ -59,7 +59,7 @@ pub fn correlated_ot_receiver(
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::ecdsa::triples::batch_random_ot::run_batch_random_ot;
+    use crate::ecdsa::ot_based_ecdsa::triples::batch_random_ot::run_batch_random_ot;
     use crate::protocol::internal::{make_protocol, Comms};
     use crate::protocol::{run_two_party_protocol, Participant};
     use k256::Secp256k1;
