+++ src/ecdsa/ot_based_ecdsa/triples/random_ot_extension.rs
@@ -243,7 +243,7 @@ fn run_random_ot<C: CSCurve>(
 
 #[cfg(test)]
 mod test {
-    use crate::ecdsa::triples::batch_random_ot::run_batch_random_ot;
+    use crate::ecdsa::ot_based_ecdsa::triples::batch_random_ot::run_batch_random_ot;
 
     use super::*;
